section_1/requirements.txt
requests
pandas
sqlalchemy
python-dotenv
tqdm
pydantic